+++ libs/framework-core/source/ftrack_framework_core/engine/__init__.py
@@ -117,6 +117,12 @@ class BaseEngine(object):
         *options*: options to be passed to the plugin
         *reference*: reference id of the plugin
         '''
+        # Specifically checking if its False, as we want to skip the plugin in that case.
+        if options.get('enabled') == False:
+            self.logger.debug(
+                f"Plugin {plugin} is disabled, skipping execution."
+            )
+            return
         registered_plugin = self.plugin_registry.get_one(name=plugin)
 
         plugin_instance = registered_plugin['extension'](
@@ -196,14 +202,18 @@ class BaseEngine(object):
 
         store = self.get_store()
         for item in engine:
+            tool_config_options = user_options.get('options') or {}
             # If plugin is just string execute plugin with no options
             if isinstance(item, str):
-                self.run_plugin(item, store, {})
+                self.run_plugin(item, store, tool_config_options)
 
             elif isinstance(item, dict):
+                if item.get("enabled") == False:
+                    continue
                 # If it's a group, execute all plugins from the group
                 if item["type"] == "group":
-                    group_options = item.get("options") or {}
+                    group_options = copy.deepcopy(tool_config_options)
+                    group_options.update(item.get("options") or {})
                     group_reference = item['reference']
                     group_options.update(
                         user_options.get(group_reference) or {}
@@ -232,10 +242,11 @@ class BaseEngine(object):
                         #  group recursively execute plugins inside
 
                 elif item["type"] == "plugin":
-                    # Execute plugin only with its own options if plugin is
-                    # defined outside the group
+                    options = copy.deepcopy(tool_config_options)
+                    options.update(item.get("options") or {})
+                    # Execute plugin only with its own options and tool_config
+                    # options if plugin is defined outside the group
                     plugin_reference = item['reference']
-                    options = item.get("options", {})
                     options.update(user_options.get(plugin_reference) or {})
                     self.run_plugin(
                         item["plugin"], store, options, plugin_reference