+++ libs/framework-core/source/ftrack_framework_core/widget/dialog.py
@@ -83,6 +83,16 @@ class FrameworkDialog(BaseUI):
 
         self._tool_config = value
 
+        if self._tool_config:
+            arguments = {
+                "tool_config_reference": self._tool_config['reference'],
+                "item_reference": None,
+                "options": self.dialog_options,
+            }
+            self.client_method_connection(
+                'set_config_options', arguments=arguments
+            )
+
         # Call _on_tool_config_changed_callback to let the UI know that a new
         # tool_config has been set.
         self._on_tool_config_changed_callback()
@@ -461,14 +471,25 @@ class FrameworkDialog(BaseUI):
         '''
         Pass the given *options* of the *plugin_reference* to the client.
         '''
+        self.set_tool_config_option(options, plugin_reference)
+
+    def set_tool_config_option(self, options, item_reference=None):
+        '''
+        Set the given name and value as options for the current tool config.
+        '''
         arguments = {
             "tool_config_reference": self.tool_config['reference'],
-            "plugin_config_reference": plugin_reference,
-            "plugin_options": options,
+            "options": options,
         }
-        self.client_method_connection(
-            'set_config_options', arguments=arguments
-        )
+        if item_reference:
+            arguments['item_reference'] = item_reference
+        self.set_option_callback(arguments)
+
+    def set_option_callback(self, args):
+        '''
+        Pass the given *args* to the client set_config_options method.
+        '''
+        self.client_method_connection('set_config_options', arguments=args)
 
     def _on_run_ui_hook_callback(self, plugin_reference, payload):
         arguments = {