+++ projects/framework-common-extensions/widgets/flie_export_options.py
@@ -55,7 +55,7 @@ class FileExportOptionsWidget(BaseWidget):
     def build_ui(self):
         '''build function widgets.'''
         # Create options:
-        for option, value in self.plugin_config.get('options').items():
+        for option, value in self.plugin_config.get('options', {}).items():
             h_layout = QtWidgets.QHBoxLayout()
             option_widget = QtWidgets.QLabel(option)
             value_widget = None