+++ projects/framework-maya/source/ftrack_framework_maya/__init__.py
@@ -9,8 +9,6 @@ import platform
 import maya.cmds as cmds
 import maya.mel as mm
 
-import ftrack_api
-
 from ftrack_framework_core.host import Host
 from ftrack_framework_core.event import EventManager
 from ftrack_framework_core.client import Client
@@ -23,6 +21,8 @@ from ftrack_utils.extensions.environment import (
     get_extensions_path_from_environment,
 )
 from ftrack_utils.usage import set_usage_tracker, UsageTracker
+
+from ftrack_utils.session import create_api_session
 
 from ftrack_framework_maya.utils import dock_maya_right, run_in_main_thread
 
@@ -92,7 +92,11 @@ def get_ftrack_menu(menu_name='ftrack', submenu_name=None):
 # mechanisms directly.
 @run_in_main_thread
 def on_run_tool_callback(
-    client_instance, tool_name, dialog_name=None, options=dict, maya_args=None
+    client_instance,
+    tool_name,
+    dialog_name=None,
+    options=None,
+    maya_args=None,
 ):
     client_instance.run_tool(
         tool_name,
@@ -111,7 +115,7 @@ def bootstrap_integration(framework_extensions_path):
         f' {framework_extensions_path}'
     )
     # Create ftrack session and instantiate event manager
-    session = ftrack_api.Session(auto_connect_event_hub=False)
+    session = create_api_session(auto_connect_event_hub=True)
     event_manager = EventManager(
         session=session, mode=constants.event.LOCAL_EVENT_MODE
     )
@@ -166,8 +170,16 @@ def bootstrap_integration(framework_extensions_path):
     )
 
     # Instantiate Host and Client
-    Host(event_manager, registry=registry_instance)
-    client_instance = Client(event_manager, registry=registry_instance)
+    Host(
+        event_manager,
+        registry=registry_instance,
+        run_in_main_thread_wrapper=run_in_main_thread,
+    )
+    client_instance = Client(
+        event_manager,
+        registry=registry_instance,
+        run_in_main_thread_wrapper=run_in_main_thread,
+    )
 
     # Init tools
     dcc_config = registry_instance.get_one(