+++ projects/framework-nuke/source/ftrack_framework_nuke/__init__.py
@@ -9,8 +9,6 @@ import platform
 
 import nuke, nukescripts
 
-import ftrack_api
-
 from ftrack_constants import framework as constants
 from ftrack_utils.extensions.environment import (
     get_extensions_path_from_environment,
@@ -23,6 +21,7 @@ from ftrack_framework_core import registry
 from ftrack_framework_core.configure_logging import configure_logging
 
 from ftrack_utils.usage import set_usage_tracker, UsageTracker
+from ftrack_utils.session import create_api_session
 
 from ftrack_framework_nuke.utils import (
     dock_nuke_right,
@@ -73,7 +72,7 @@ startup_tools = []
 
 
 @run_in_main_thread
-def on_run_tool_callback(tool_name, dialog_name=None, options=dict):
+def on_run_tool_callback(tool_name, dialog_name=None, options=None):
     client_instance.run_tool(
         tool_name,
         dialog_name,
@@ -93,7 +92,7 @@ def bootstrap_integration(framework_extensions_path):
         f' {framework_extensions_path}'
     )
 
-    session = ftrack_api.Session(auto_connect_event_hub=False)
+    session = create_api_session(auto_connect_event_hub=True)
 
     event_manager = EventManager(
         session=session, mode=constants.event.LOCAL_EVENT_MODE
@@ -147,9 +146,16 @@ def bootstrap_integration(framework_extensions_path):
         )
     )
 
-    Host(event_manager, registry=registry_instance)
-
-    client_instance = Client(event_manager, registry=registry_instance)
+    Host(
+        event_manager,
+        registry=registry_instance,
+        run_in_main_thread_wrapper=run_in_main_thread,
+    )
+    client_instance = Client(
+        event_manager,
+        registry=registry_instance,
+        run_in_main_thread_wrapper=run_in_main_thread,
+    )
 
     # Init tools
     dcc_config = registry_instance.get_one(
@@ -167,7 +173,7 @@ def bootstrap_integration(framework_extensions_path):
         on_menu = tool.get("menu", True)
         name = tool['name']
         dialog_name = tool.get('dialog_name')
-        options = tool.get('options')
+        options = tool.get('options', {})
         # TODO: In the future, we should probably emit an event so plugins can
         #  subscribe to it. and run_on specific event.
         if on_menu: