+++ projects/framework-houdini/release_notes.md
@@ -1,5 +1,13 @@
 # ftrack Framework Houdini integration release Notes
 
+
+## upcoming
+
+* [changed] Init; Use create_api_session utility to create the api session.
+* [changed] Host, Client instance; Pass run_in_main_thread argument.
+* [fix] Init; Fix on_run_tool_callback options argument.
+
+
 ## v24.6.0
 2024-06-26