+++ projects/framework-houdini/release_notes.md
@@ -6,6 +6,7 @@
 * [changed] Init; Use create_api_session utility to create the api session.
 * [changed] Host, Client instance; Pass run_in_main_thread argument.
 * [fix] Init; Fix on_run_tool_callback options argument.
+* [fix] Adjust version expression so Houdini Launchers can be created properly.
 
 
 ## v24.6.0