+++ projects/framework-houdini/release_notes.md
@@ -1,8 +1,12 @@
 # ftrack Framework Houdini integration release Notes
 
 
-## upcoming
+## v24.10.0
+2024-10-21
 
+* [changed] Build plugin with python 3.10 instead of 3.9
+* [changed] Update library versions from version 2 to 3
+* [fix] Adjust version expression so Houdini Launchers can be created properly.
 * [changed] Init; Use create_api_session utility to create the api session.
 * [changed] Host, Client instance; Pass run_in_main_thread argument.
 * [fix] Init; Fix on_run_tool_callback options argument.